# src/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Todo App",
    "description": "Documentation for Todo App"
  },
  "host": "localhost:5000",
  "paths": {
    "/users": {
      "get": {
        "summary": "Getting all users",
        "description": "Method for getting all users",
        "operationId": "users",
        "tags": ["Users"],
        "responses": {
          "200": {
            "description": "Users getting was successful"
          }
        }
      },
      "post": {
        "summary": "User creation",
        "description": "Method for creating a user",
        "operationId": "createUser",
        "produces": ["application/json"],
        "tags": ["Users"],
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "Create user",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewUser"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User was successfully created",
            "schema": {
              "$ref": "#/definitions/NewUser"
            }
          }
        }
      }
    },
...
    "/users/{userId}/lists/{listId}/todoItems": {
      "get": {
        "summary": "Getting all todo items of certain todo list",
        "description": "Method for getting all todo items by todo id",
        "tags": ["Todo items"],
        "operationId": "items",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "Id of user",
            "required": true,
            "type": "integer"
          },
          {
            "name": "listId",
            "in": "path",
            "description": "Id of todo list",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Todo items getting was successful"
          }
        }
      },
      "post": {
        "summary": "Todo item creation",
        "description": "Method for creating a todo item",
        "tags": ["Todo items"],
        "operationId": "createTodoItem",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Todo Item",
            "in": "body",
            "description": "Create todo item",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewTodoItem"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Todo item was successfully created",
            "schema": {
              "$ref": "#/definitions/NewTodoItem"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "NewUser": {
      "type": "object",
      "required": ["firstName", "lastName", "email", "password", "role"],
      "properties": {
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "role": {
          "type": "string"
        }
      }
    },
...
    "NewTodoItem": {
      "type": "object",
      "required": ["user_id", "todo_id", "todoTitle", "isCompleted"],
      "properties": {
        "userId": {
          "type": "integer"
        },
        "todoId": {
          "type": "integer"
        },
        "todoTitle": {
          "type": "string"
        },
        "isCompleted": {
          "type": "boolean"
        }
      }
    }
  }
}